isConnected should check for local endpoints using path
Thanks to @krishnasrinivas for finding this issue, current
issue was that we match storageAPI.String() returns the disk
path and endpoint.String() still returns full endpoint URL
when disks are local. This matching error leads to isConnected
to return false. The side affect of this is such that we will
be continually replacing the local disks with new StorageAPI.